requirements.txt
matplotlib==3.5.3
numpy==1.21.6
ogb==1.3.3
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.7.3
torch-geometric==2.0.4
torch-scatter @ https://data.pyg.org/whl/torch-1.9.0%2Bcu111/torch_scatter-2.0.9-cp37-cp37m-linux_x86_64.whl
torch-sparse @ https://data.pyg.org/whl/torch-1.9.0%2Bcu111/torch_sparse-0.6.12-cp37-cp37m-linux_x86_64.whl
tqdm==4.64.1
numba==0.56.4
